+++ src/schemas/users/StatisticsSchema.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod'
 
 export const UserStatisticSchema = z.object({
-  views: z.number(),
-  likes: z.number(),
-  followers: z.number(),
+  views: z.number().int(),
+  likes: z.number().int(),
+  followers: z.number().int(),
 })
 
 export type UserStatisticSchemaType = z.infer<typeof UserStatisticSchema>
